+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Emoji 12.0 data
 - Introduce new `REGEX_WELL_FORMED` to be able to match for invalid tag and region sequences
+- Introduce new `*_INCLUDE_TEXT` regexes which include matching for textual presentation emoji
 - Refactoring: Update Emoji matching to latest standard while keeping naming close to standard
 - Issue warning when using `#list` method to retrieve outdated category
 - Change matching for ZWJ sequences: Do not limit sequence to a maximum of 3 ZWJs